Extract tabol into separate file
